// src/api/controllers/product/product.controller.d.ts
import type { Request, Response } from "express";

export interface ProductController {
  getAll: (req: Request, res: Response) => Promise<void>;
  getOne: (req: Request, res: Response) => Promise<void>;
  createOne: (req: Request, res: Response) => Promise<void>;
  updateOne: (req: Request, res: Response) => Promise<void>;
  deleteOne: (req: Request, res: Response) => Promise<void>;
}